requirements.txt
absl-py==1.2.0
accelerate==0.12.0
aiohttp==3.8.3
aiosignal==1.2.0
ale-py==0.8.0
apex==0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asynctest==0.13.0
atari-py==0.2.6
attrs==22.1.0
beautifulsoup4==4.11.1
bleach==5.0.1
box2d-py==2.3.8
Brotli==1.0.9
cachetools==4.2.4
catboost==1.1.1
cattrs==1.0.0
cffi==1.15.1
charset-normalizer==2.1.0
chess==1.7.0
click==8.1.3
cloudpickle==1.3.0
cycler==0.11.0
Cython==0.29.32
datasets==2.5.1
decorator==4.4.2
deepspeed==0.7.3
defusedxml==0.7.1
diffusers==0.10.2
dill==0.3.5.1
docker-pycreds==0.4.0
docopt==0.6.2
EasyProcess==1.1
einops==0.6.0
ema-pytorch==0.0.10
evaluate==0.2.2
fasteners==0.15
fastjsonschema==2.16.2
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.48.0
gym==0.18.3
gym-notices==0.0.8
h5py==3.8.0
hanabi-learning-environment==0.0.4
hjson==3.1.0
huggingface-hub==0.11.1
idna==3.3
imageio==2.5.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflate64==0.3.0
ipdb==0.13.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaggle==1.5.12
kiwisolver==1.4.4
kornia==0.6.10
lightgbm==3.3.3
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mistune==2.0.4
monotonic==1.6
moviepy==1.0.0
mujoco-py==2.1.2.14
multidict==6.0.2
multiprocess==0.70.13
multivolumefile==0.2.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
networkx==2.5
ninja==1.10.2.3
nltk==3.7
notebook==6.4.12
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.4.0.42
pandas==1.3.5
pandocfilters==1.5.0
pathtools==0.1.2
PettingZoo==1.0.0
Pillow==8.2.0
pkgutil_resolve_name==1.3.10
plotly==5.11.0
portalocker==2.5.1
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.0
py-cpuinfo==8.0.0
py7zr==0.20.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.2
pygame==2.1.0
pyglet==1.5.0
Pygments==2.13.0
pyppmd==0.18.3
pyrsistent==0.18.1
python-slugify==6.1.2
pytorch-fid==0.2.1
pytz==2022.2.1
PyVirtualDisplay==1.3.2
PyYAML==6.0
pyzstd==0.15.3
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rlcard==1.0.5
rsa==4.9
sacrebleu==2.2.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.17.0
setproctitle==1.3.2
simplegeneric==0.8.1
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tenacity==8.1.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
tensorboardX==1.8
termcolor==2.0.1
terminado==0.15.0
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.14.0
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wincertstore==0.2
xgboost==1.6.2
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
